requirements.txt
blinker==1.4
flask==2.0.3
flask-dance==6.0.0
flask-jwt-extended==4.4.1
flask-login==0.6.1
flask-marshmallow==0.14.0
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
gunicorn==20.1.0
marshmallow-sqlalchemy==0.28.0
pandas==1.3.5
psycopg2-binary==2.9.3
python-dotenv==0.20.0
setuptools==62.1.0
sqlalchemy-utils==0.38.2
